# OpenZFS 8199 - multi-threaded dmu_object_alloc()