@ahrens 2017-05-11
Thanks @behlendorf.  I updated the comment as suggested.