@ahrens 2017-05-24
@nedbass I think I've addressed your feedback, could you take one last look?  Also could you let me know if the failed tests are something I need to look at?  I clicked around a bit and didn't find any errors that I could understand :-)