@nedbass 2017-05-24
@ahrens, thanks for addressing the feedback. We had also proposed renaming `os->os_obj_next` to `os->os_obj_next_chunk`. If you still agree with that naming choice then that is still unaddressed.

Regarding the failures, it looks like we tripped the assert `ASSERT0(P2PHASE(os->os_obj_next, dnodes_per_chunk));`

```
[ 2328.090441] SPL: using hostid 0x00000000
[ 2328.134694] VERIFY3(0 == ((os->os_obj_next) & ((dnodes_per_chunk) - 1))) failed (0 == 2)
[ 2328.141374] PANIC at dmu_object.c:93:dmu_object_alloc_dnsize()
[ 2328.146082] Showing stack for process 30735
[ 2328.150109] CPU: 0 PID: 30735 Comm: zpool Tainted: P           OE  ------------   3.10.0-514.16.1.el7.x86_64 #1
[ 2328.157723] Hardware name: Xen HVM domU, BIOS 4.2.amazon 11/11/2016
[ 2328.161625]  ffffffffa07b5104 000000006cb6831b ffff8801bb66b8f8 ffffffff81686ac3
[ 2328.166631]  ffff8801bb66b908 ffffffffa04b2e34 ffff8801bb66ba90 ffffffffa04b2eff
[ 2328.171484]  000000006cb6831b ffff880100000030 ffff8801bb66baa0 ffff8801bb66ba40
[ 2328.176748] Call Trace:
[ 2328.179740]  [<ffffffff81686ac3>] dump_stack+0x19/0x1b
[ 2328.185527]  [<ffffffffa04b2e34>] spl_dumpstack+0x44/0x50 [spl]
[ 2328.191503]  [<ffffffffa04b2eff>] spl_panic+0xbf/0xf0 [spl]
[ 2328.196223]  [<ffffffff81698eb1>] ? ftrace_call+0x5/0x2f
[ 2328.200263]  [<ffffffff8168e305>] ? _raw_spin_unlock+0x5/0x30
[ 2328.206245]  [<ffffffffa0621b78>] ? dbuf_rele_and_unlock+0x1f8/0x5c0 [zfs]
[ 2328.211998]  [<ffffffffa063437f>] dmu_object_alloc_dnsize+0x2af/0x470 [zfs]
[ 2328.217338]  [<ffffffffa06c954b>] zap_create_norm_dnsize+0x2b/0x60 [zfs]
[ 2328.225800]  [<ffffffffa06c95f2>] zap_create+0x22/0x30 [zfs]
[ 2328.230656]  [<ffffffffa06fb1fa>] zfs_create_fs+0x84a/0xa90 [zfs]
[ 2328.235431]  [<ffffffff81698eb1>] ? ftrace_call+0x5/0x2f
[ 2328.240017]  [<ffffffff81698eb1>] ? ftrace_call+0x5/0x2f
[ 2328.244874]  [<ffffffffa0668a33>] dsl_pool_create+0x1e3/0x500 [zfs]
[ 2328.250504]  [<ffffffffa0692aef>] spa_create+0x5af/0xb30 [zfs]
[ 2328.254597]  [<ffffffffa06e0862>] zfs_ioc_pool_create+0x152/0x270 [zfs]
[ 2328.258937]  [<ffffffffa06e0715>] ? zfs_ioc_pool_create+0x5/0x270 [zfs]
[ 2328.262938]  [<ffffffffa06e16a7>] zfsdev_ioctl+0x537/0x660 [zfs]
[ 2328.267654]  [<ffffffffa06e1170>] ? pool_status_check+0xf0/0xf0 [zfs]
[ 2328.272678]  [<ffffffff81212555>] do_vfs_ioctl+0x2d5/0x4b0
[ 2328.276149]  [<ffffffff812127d1>] SyS_ioctl+0xa1/0xc0
[ 2328.279415]  [<ffffffff81697189>] system_call_fastpath+0x16/0x1b
[ 2520.231124] INFO: task zpool:30735 blocked for more than 120 seconds.
```