@ahrens 2017-05-24
@nedbass Sounds good I'll change it to `os_obj_next_chunk`.

Ah, I must have missed this link, I assume that's where you found the stack trace:
http://build.zfsonlinux.org/builders/CentOS%207%20x86_64%20Mainline%20%28TEST%29/builds/2504/steps/shell_6/logs/console
I'll investigate that.