@nedbass 2017-05-24
@ahrens Yes, check the console log anytime you see tests fails due to timeout, as that is where you'll find any failed ASSERTs and kernel stack traces.