@ahrens 2017-05-27
@nedbass I think there are two ways that this assertion can fail:

The `restarted` logic could make us find the next block of dnodes, rather than the next L1.  So the alignment of `os_obj_next` could be only a single block’s worth (32) — which is less than `dmu_object_alloc_chunk_shift` (which is between a block’s worth an a L1's worth).

`zfs_create_fs()` modifies `os_obj_next`, setting it to exactly `2`.  Looks like that change was also part of the large blocks project.

Any advice on how to handle either of these?