@nedbass 2017-05-31
> The restarted logic could make us find the next block of dnodes, rather than the next L1. So the alignment of os_obj_next could be only a single block’s worth (32) — which is less than dmu_object_alloc_chunk_shift (which is between a block’s worth an a L1's worth).

My thought here is to just mask off the lower bits when assigning `os_obj_next_chunk` so it's always a multiple of `dnodes_per_chunk`.

> zfs_create_fs() modifies os_obj_next, setting it to exactly 2. Looks like that change was also part of the large dnode project.

I think we can remove the code that sets it to 2. That was needed when we were advancing `os_obj_next` to the next dnode block in case `dmu_object_next()` returned an error, which it did when `object == 0`  at fs creation time. But with your re-working of the code that case now looks to be handled by the global allocator logic.

The former consequence of not initializing `os_obj_next` to 2 was that object numbers of files in a brand-new dataset would begin at 32. If you observe new object numbers beginning < 32 (typically 7) after removing that code then it should be fine to leave it out.

This does bring up the point that there is no handling of errors from `dmu_object_next()`. What happens if it returns ESRCH?  It looks like that might result in an infinite loop. Do we think that's not possible when searching for a hole? Perhaps that should be documented in the form of an ASSERT or a comment.