@ahrens 2017-05-31
@nedbass 
> My thought here is to just mask off the lower bits when assigning os_obj_next_chunk so it's always a multiple of dnodes_per_chunk.

Will do.

> I think we can remove the code that sets it to 2. 

Will do.

> This does bring up the point that there is no handling of errors from dmu_object_next().

I like your logic of advancing to the next L0.  That way at least it can be sure to make progress, and eventually rescan if necessary.  I'll put that back in place.