@nedbass 2017-05-31
> You mentioned that dmu_object_next() returned an error when object == 0. I didn't see where in the code that would happen, so I'm having trouble understanding why it seems to be getting an error on the MOS but not on regular objects.

@ahrens, I bet it's hitting this case in `dnode_next_offset()`:

```
        if (dn->dn_phys->dn_nlevels == 0) {
                error = SET_ERROR(ESRCH);
                goto out;
        }
```

If I recall, that's why I made the comment (removed in your last commit) about the metadnode's dnode_phys_t structure not being initialized yet. Not sure what the best way to deal with that is, which is why I ended up advancing `os_obj_next` to 2. That way the initial `dnode_hold_impl()` call would succeed and thereby avoid the call to `dmu_object_next()` during dataset creation.