@ahrens 2017-06-01
@nedbass I see, the problem isn't calling dmu_object_next(object=0), but rather calling dmu_object_next() at all when the objset has not been synced to disk yet.  And this would happen when we first call dmu_object_alloc() because the call to dnode_hold_impl(object=0) would fail.

I'm still trying to figure out why this doesn't happen for ordinary filesystems, for the objects allocated from zfs_create_fs().  But I don't think it's a big deal either way.