@behlendorf 2017-06-02
@ahrens when you're happy with this go ahead and squash the commits and add your signed-off-by.  We'll queue this up for some additional local testing too.