@nedbass 2017-06-05
@ahrens  also don't forget about renaming `os_obj_next` to `os_obj_next_chunk` when you rebase :smile: