@tonyhutter 2017-06-05
@ahrens  I'm seeing ANSI codes in your commit message:
```
    The solution is to have a <E2><80><9C>next object to allocate<E2><80><9D> for each CPU. Each
    of these <E2><80><9C>next object<E2><80><9D>s is in a different block of the dnode object, so
    that concurrent allocation holds dnodes in different dbufs.  When a
    thread<E2><80><99>s <E2><80><9C>next object<E2><80><9D> reaches the end of a chunk of objects (by default
    4 blocks worth -- 128 dnodes), it will be reset to the per-objset
    os_obj_next, which will be increased by a chunk of objects (128).  
   ...
```