@tonyhutter 2017-06-05
I benchmarked this using the same test setup as https://github.com/zfsonlinux/zfs/pull/5935#issuecomment-306263532 (16 core Xeon, ramdisk pool, results averaged over 3 runs).  The numbers look great!

File creates per second:

| Threads | Before | After | Speedup |
| --- | --- | --- | --- |
| 1 | 17928.2 | 17796.7 | -0.73% |
| 2 | 29141.5 | 31927.1 | 9.56% |
| 4 | 54211.7 | 56868.7 | 4.90% |
| 8 | 85954.5 | 99901.9 | 16.23% |
| 16 | 108344.7 | 150460.7 | 38.87% |
| 32 | 112364.9 | 154242.2 | 37.27% |
| 64 | 111778.3 | 156167.0 | 39.71% |
| 128 | 113122.6 | 158316.7 | 39.95% |
| 256 | 113899.4 | 154484.2 | 35.63% |