@tonyhutter 2017-06-06
More results, this time using real hardware.  This is a SSD pool used by one of our lustre MDS nodes (16 core):
```
 	jet1        ONLINE       0     0     0
	  mirror-0  ONLINE       0     0     0
	    U0      ONLINE       0     0     0
	    U1      ONLINE       0     0     0
	    L2      ONLINE       0     0     0
	  mirror-1  ONLINE       0     0     0
	    L0      ONLINE       0     0     0
	    L1      ONLINE       0     0     0
	    U2      ONLINE       0     0     0
```
File creates/sec:

| Threads | Before | After | Speedup |
| --- | --- | --- | --- |
| 1 | 10255 | 10469 | 2% |
| 2 | 18168 | 19707 | 8% |
| 4 | 32316 | 35476 | 10% |
| 8 | 52143 | 56630 | 9% |
| 16 | 57917 | 76697 | 32% |
| 32 | 58791 | 76357 | 30% |
| 64 | 58643 | 77782 | 33% |
| 128 | 59471 | 79658 | 34% |
| 256 | 59713 | 82078 | 37% |